CM-792: Create parkCard component for find a park page
